+++ Controller/Test/Test.php
@@ -10,27 +10,23 @@ namespace Ebizmarts\SagePaySuite\Controller\Test;
 
 class Test extends \Magento\Framework\App\Action\Action
 {
-//
-//    protected $_sagepay_transaction;
-//
-//
-//    public function __construct(
-//        \Magento\Framework\App\Action\Context $context,
-//        \Ebizmarts\SagePaySuite\Model\Api\Transaction $sp_transaction
-//    ){
-//        parent::__construct($context);
-//
-//        $this->_sagepay_transaction = $sp_transaction;
-//    }
-//
-//    public function execute()
-//    {
-//
-//        try {
-//            //$result = $this->_sagepay_transaction->voidTransaction("2F1D88C1-75B6-0CB2-C042-3AF0B19AE21B");
-//        } catch (\Ebizmarts\SagePaySuite\Model\Api\ApiException $apiException) {
-//
-//            $error_message = $apiException->getUserMessage();
-//        }
-//    }
+
+    protected $cron;
+
+
+    public function __construct(
+        \Magento\Framework\App\Action\Context $context,
+        \Ebizmarts\SagePaySuite\Model\Cron $cron
+    ){
+        parent::__construct($context);
+
+        $this->cron = $cron;
+    }
+
+    public function execute()
+    {
+
+        //$this->cron->checkFraud();
+        $this->cron->cancelPendingPaymentOrders();
+    }
 }