+++ Helper/Request.php
@@ -1,0 +1,57 @@
+<?php
+/**
+ * Copyright © 2015 ebizmarts. All rights reserved.
+ * See LICENSE.txt for license details.
+ */
+
+namespace Ebizmarts\SagePaySuite\Helper;
+
+
+class Request extends \Magento\Framework\App\Helper\AbstractHelper
+{
+
+    public function populateAddressInformation($quote){
+
+        $billing_address = $quote->getBillingAddress();
+        $shipping_address = $quote->isVirtual() ? $billing_address : $quote->getShippingAddress();
+
+        $data = array();
+
+        $data['BillingSurname']    = substr($billing_address->getLastname(), 0, 20);
+        $data['BillingFirstnames'] = substr($billing_address->getFirstname(), 0, 20);
+        $data['BillingAddress1']   = substr($billing_address->getStreetLine(1), 0, 100);
+        $data['BillingCity']       = substr($billing_address->getCity(), 0,  40);
+        $data['BillingPostCode']   = substr($billing_address->getPostcode(), 0, 10);
+        $data['BillingCountry']    = substr($billing_address->getCountryId(), 0, 2);
+
+        //only send state if US due to Sage Pay 2 char restriction
+        if ($data['BillingCountry'] == 'US') {
+            $data['BillingState'] = substr($billing_address->getRegionCode(), 0, 2);
+        }
+
+        //not mandatory
+//        $data['BillingAddress2']   = ($this->getConfigData('mode') == 'test') ? 88 : $this->ss($billing->getStreet(2), 100);
+
+
+        //mandatory
+        $data['DeliverySurname']    = substr($shipping_address->getLastname(), 0, 20);
+        $data['DeliveryFirstnames'] = substr($shipping_address->getFirstname(), 0, 20);
+        $data['DeliveryAddress1']   = substr($shipping_address->getStreetLine(1), 0, 100);
+        $data['DeliveryCity']       = substr($shipping_address->getCity(), 0,  40);
+        $data['DeliveryPostCode']   = substr($shipping_address->getPostcode(), 0, 10);
+        $data['DeliveryCountry']    = substr($shipping_address->getCountryId(), 0, 2);
+        //only send state if US due to Sage Pay 2 char restriction
+        if ($data['DeliveryCountry'] == 'US') {
+            $data['DeliveryState'] = substr($shipping_address->getRegionCode(), 0, 2);
+        }
+
+        //not mandatory
+//        $data['DeliveryAddress2']   = ($this->getConfigData('mode') == 'test') ? 88 : $this->ss($billing->getStreet(2), 100);
+//        $data['DeliveryState'] = $billing->getRegionCode();
+//        $data['DeliveryPhone'] = $billing->getRegionCode();
+
+        return $data;
+
+    }
+
+}